Correct False Update Trigger
Oxide was triggering an update when it shouldn't. There was two related reasons. The first appeared to be the server not starting in time for the new version to be written to the oxide logs before the update check ran again. I added an extra 20 minute wait before checking again, plus the wait set by the user. This should give the server time to start and write the logs. It's pretty unlikely an update would come down this quickly.

Second was due to the oxide logs rolling over nightly. I modified the log search to check through the last two log files in the directory to make sure the update isn't triggered while the new log is being written.