Complete Basic Changes for Seed Wipes
Made and tested basic changes to allow detecting the first thursday of the month, and then updating the seed if it has been at least 27 days ( need to make this configurable ). The seed will be logged, and the configuration will be written with the new seed. Probably need to make this do a backup of the configuration and seed log just in case.