Modified Logging Setup and File Backup
Modified logging setup to handle multiple calls to the class. Added function to change the log level without having to call a new class.

Changed File Backup to throw error if file not found after backup and to not throw an error if unable to remove old backup.

Added logging to certain places, and added backup to Configuration File before writing new one.