Backups before writing files in mapseed and regex improvements
Added option for backup to be passed full filepath including the name in filepath by passing empty string to filename.  Added backup options to mapseed and adjusted the regex on the seed search to pick up most (I think) possibilities of how the seed could be formatted in the configuration.